+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `is_socks` function
 ### Changed
 - `README.md` updated
+- Test system modified
 ## [0.2] - 2026-06-04
 ### Added
 - `VMESSConfig` class